+++ cli/src/commands/diff.rs
@@ -29,10 +29,10 @@ use crate::cli_util::print_unmatched_explicit_paths;
 use crate::cli_util::short_commit_hash;
 use crate::command_error::CommandError;
 use crate::command_error::user_error;
+use crate::commit_templater::TreeDiff;
 use crate::complete;
 use crate::diff_util::DiffFormatArgs;
 use crate::diff_util::get_copy_records;
-use crate::diff_util::show_templated;
 use crate::ui::Ui;
 
 /// Compare file contents between two revisions
@@ -90,12 +90,12 @@ pub(crate) struct DiffArgs {
 
     /// Render each file diff entry using the given template
     ///
-    /// All 0-argument methods of the [`TreeDiffEntry` type] are available as
+    /// All 0-argument methods of the [`TreeDiff` type] are available as
     /// keywords in the template expression. See [`jj help -k templates`] for
     /// more information.
     ///
-    /// [`TreeDiffEntry` type]:
-    ///     https://docs.jj-vcs.dev/latest/templates/#treediffentry-type
+    /// [`TreeDiff` type]:
+    ///     https://docs.jj-vcs.dev/latest/templates/#treediff-type
     ///
     /// [`jj help -k templates`]:
     ///     https://docs.jj-vcs.dev/latest/templates/
@@ -183,38 +183,39 @@ pub(crate) fn cmd_diff(
         }
     }
 
-    // -T disables both short/long rendering formats, but it might be okay to
-    // enable long format if explicitly specified (assuming -T is for short or
+    // -T disables emitting other formats, but it might be okay to emit
+    // both if explicitly specified (assuming -T is for short or
     // summary output.)
-    let maybe_template;
-    let diff_renderer;
+
+    ui.request_pager();
+
     if let Some(text) = &args.template {
         let language = workspace_command.commit_template_language();
         let template = workspace_command
             .parse_template(ui, &language, text)?
             .labeled(["diff"]);
-        maybe_template = Some(template);
-        diff_renderer = workspace_command.diff_renderer(vec![]);
+        let tree_diff = TreeDiff::new(
+            from_tree.clone(),
+            to_tree.clone(),
+            matcher.into(),
+            copy_records,
+        );
+        template.format(&tree_diff, ui.stdout_formatter().as_mut())?;
     } else {
-        maybe_template = None;
-        diff_renderer = workspace_command.diff_renderer_for(&args.format)?;
+        let diff_renderer = workspace_command.diff_renderer_for(&args.format)?;
+
+        diff_renderer
+            .show_diff(
+                ui,
+                ui.stdout_formatter().as_mut(),
+                Diff::new(&from_tree, &to_tree),
+                &matcher,
+                &copy_records,
+                ui.term_width(),
+            )
+            .block_on()?;
     }
 
-    ui.request_pager();
-    if let Some(template) = &maybe_template {
-        let tree_diff = from_tree.diff_stream_with_copies(&to_tree, &matcher, &copy_records);
-        show_templated(ui.stdout_formatter().as_mut(), tree_diff, template).block_on()?;
-    }
-    diff_renderer
-        .show_diff(
-            ui,
-            ui.stdout_formatter().as_mut(),
-            Diff::new(&from_tree, &to_tree),
-            &matcher,
-            &copy_records,
-            ui.term_width(),
-        )
-        .block_on()?;
     print_unmatched_explicit_paths(
         ui,
         &workspace_command,