+++ cli/src/diff_util.rs
@@ -84,7 +84,6 @@ use unicode_width::UnicodeWidthStr as _;
 
 use crate::command_error::CommandError;
 use crate::command_error::cli_error;
-use crate::commit_templater;
 use crate::config::CommandNameAndArgs;
 use crate::formatter::Formatter;
 use crate::formatter::FormatterExt as _;
@@ -95,7 +94,6 @@ use crate::merge_tools::ExternalMergeTool;
 use crate::merge_tools::generate_diff;
 use crate::merge_tools::invoke_external_diff;
 use crate::merge_tools::new_utf8_temp_dir;
-use crate::templater::TemplateRenderer;
 use crate::text_util;
 use crate::ui::Ui;
 
@@ -2208,18 +2206,6 @@ pub async fn show_names(
             "{}",
             path_converter.format_file_path(path.target())
         )?;
-    }
-    Ok(())
-}
-
-pub async fn show_templated(
-    formatter: &mut dyn Formatter,
-    mut tree_diff: BoxStream<'_, CopiesTreeDiffEntry>,
-    template: &TemplateRenderer<'_, commit_templater::TreeDiffEntry>,
-) -> Result<(), DiffRenderError> {
-    while let Some(entry) = tree_diff.next().await {
-        let entry = commit_templater::TreeDiffEntry::from_backend_entry_with_copies(entry)?;
-        template.format(&entry, formatter)?;
     }
     Ok(())
 }